Moved partials from `views.py` to `views_partials.py`
