+++ hypha/apply/funds/urls.py
@@ -33,8 +33,6 @@ from .views import (
     UpdatePartnersView,
     UpdateReviewersView,
     htmx_archive_unarchive_submission,
-    partial_screening_card,
-    partial_translate_answers,
     reminder_list,
     submission_success,
 )
@@ -49,9 +47,11 @@ from .views_partials import (
     partial_meta_terms_card,
     partial_reviews_card,
     partial_reviews_decisions,
+    partial_screening_card,
     partial_submission_activities,
     partial_submission_answers,
     partial_submission_lead,
+    partial_translate_answers,
     sub_menu_bulk_update_lead,
     sub_menu_bulk_update_reviewers,
     sub_menu_category_options,
